// src/types/progress.ts
export type Progression =
  | 'Topic Approval'
  | 'Adviser Invitation'
  | 'Chapt 1-3'
  | 'Chapt 1-3 Approval'
  | 'Proposal Paper Forms'
  | 'Proposal Defense'
  | 'Compliance Matrix (Proposal)'
  | 'Revisions Chapt 1-3'
  | 'System Development'
  | 'Compliance Matrix Approval (Proposal)'
  | 'Chapt 4-5'
  | 'Oral Defense Form'
  | 'Oral Defense'
  | 'Compliance Matrix (Oral)'
  | 'Capstone Paper Revisions'
  | 'Compliance Matrix Approval (Oral)'
  | 'Library Hardbound';